+++ CHANGELOG.md
@@ -42,7 +42,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Fixed some issues with IPv6 ASes
   - New option `EXPERIMENTAL_SCION_RESOLVER_MINIMIZE_REQUESTS`
 - "Integration" test for scionproto "default". 
-  [#116](https://github.com/scionproto-contrib/jpan/pull/116)
+  [#114](https://github.com/scionproto-contrib/jpan/pull/114)
+- Improved path duplication filtering.
+  [#117](https://github.com/scionproto-contrib/jpan/pull/117)
 
 ### Changed
 - Clean up TODO and deprecation info. [#100](https://github.com/scionproto-contrib/jpan/pull/100)